Refactor and optimize various components in the kernel
- Introduced a new `merge_pass` function in `DirtyRegionManager` to streamline the merging of dirty rectangles.
- Refactored `mark_dirty` in `Framebuffer` to utilize `try_merge_or_insert_dirty` for better clarity and efficiency.
- Extracted alias collection logic into a separate function in `AmdIommuDriver` for improved readability.
- Simplified `commit_allocated_tables` in `IommuDomain` by removing redundant function definition.
- Enhanced `try_map_rmrr_region` in `dma.rs` to directly use `start` and `size` as `u64` for consistency.
- Updated `verify_no_reserved_slots` to accept `QuarantineQueueInner` for better type clarity.
- Changed `sg` parameter type from `u64` to `u32` in `process_single_zombie` for consistency in `ZombieQueue`.
- Improved interrupt handling in `VirtioConsoleDevice` by separating TX and RX processing into distinct functions.
- Added a new validation function for memory descriptors in `memory.rs` to enhance memory management.
- Refactored `split_and_reinsert_regions` in `ProcessAddressSpace` to work with `MemoryRegion` instead of `AddressRegion`.
- Optimized `find_contiguous_set_bits` in `buddy_allocator.rs` for better performance and clarity.
- Simplified NUMA node initialization in `NumaPmmAllocator` by creating a dedicated function for node-specific setup.
- Introduced a utility function to check if a range is within NUMA regions in `frame_allocator.rs`.
- Adjusted PAT flag handling for huge pages in `higher_half.rs` to improve clarity.
- Removed redundant merging logic in `RemoteFreeRing` to streamline entry processing.
- Enhanced safety in `SlabCache` by ensuring deallocation is performed safely.
- Improved ARP cache management by introducing a dedicated function to find available slots.
- Added a new method to parse certificate request extensions in `TlsConnection` for better TLS handling.
- Refactored `ExprParser` to simplify else branch parsing logic.
- Enhanced task searching logic in `Executor` to improve efficiency and clarity.
- Improved fan control logic in `FanController` by mapping temperature to fan levels in a dedicated function.
- Simplified corner cutout logic in `IconGenerator` for better readability and maintainability.

+++ kernel/src/io/iommu/intel/controller/dma.rs
@@ -133,10 +133,10 @@ fn validate_rmrr_region(start: u64, end: u64) -> Result<(), IommuError> {
 fn try_map_rmrr_region(
     domain: &IommuDomain,
     device: DeviceId,
-    start: usize,
-    size: usize,
+    start: u64,
+    size: u64,
 ) -> Result<(), IommuError> {
-    if let Err(e) = validate_rmrr_region(start as u64, (start + size) as u64) {
+    if let Err(e) = validate_rmrr_region(start, start + size) {
         log::error!(
             "[IOMMU][CRITICAL] RMRR validation failed for {:04x}:{:02x}:{:02x}.{}: \
              region {:#x}-{:#x}, error: {:?}",
@@ -146,7 +146,7 @@ fn try_map_rmrr_region(
         return Err(IommuError::RmrrMapFailed);
     }
 
-    match domain.map(start as u64, start as u64, size as u64, true, true) {
+    match domain.map(start, start, size, true, true) {
         Ok(()) => {
             log::debug!(
                 "[IOMMU] RMRR mapped for {:04x}:{:02x}:{:02x}.{}: {:#x}-{:#x}",
